requirements.txt
# SehatAgent - Requirements
# Multi-Agent AI Health System for IDEAX92
# Windows-compatible version

# ============ Core Framework ============
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============ Google Cloud ============
google-cloud-aiplatform>=1.38.0
google-cloud-speech>=2.23.0
google-cloud-storage>=2.14.0

# ============ Database ============
sqlalchemy>=2.0.25
asyncpg>=0.29.0
greenlet>=3.0.0

# PostgreSQL driver - psycopg3 works on Windows without pg_config
psycopg[binary]>=3.1.0

# ============ Vector Database (FAISS) ============
faiss-cpu>=1.7.0
numpy>=1.26.0

# ============ Language Processing ============
langdetect>=1.0.9
deep-translator>=1.11.4

# ============ Audio Processing ============
pydub>=0.25.1
soundfile>=0.12.1

# ============ HTTP & Async ============
httpx>=0.26.0
aiofiles>=23.2.1
tenacity>=8.2.3

# ============ Logging & Monitoring ============
structlog>=24.1.0

# ============ Utilities ============
python-dotenv>=1.0.0

# ============ Data Processing ============
pandas>=2.1.0
